pyproject.toml
[project]
name = "testmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
]
